create compact from existing context, adapt pred/succ testing to CompactContext

+++ memberships/src/main/java/com/salesforce/apollo/membership/Context.java
@@ -231,7 +231,7 @@ public interface Context<T extends Member> {
     /**
      * Answer the number of rings in the context
      */
-    int getRingCount();
+    short getRingCount();
 
     Digest hashFor(Digest d, int ring);
